Complete modularized of all commands
